+++ bin/deimos/Cargo.toml
@@ -1,0 +1,2 @@
+[package]
+name = "deimos"
