+++ crates/api/src/config.rs
@@ -1,0 +1,16 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct WebServerConfig {
+    pub host: String,
+    pub port: u16,
+}
+
+impl Default for WebServerConfig {
+    fn default() -> Self {
+        WebServerConfig {
+            host: "127.0.0.1".to_string(),
+            port: 8080,
+        }
+    }
+}
