+++ crates/api/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod config;
+pub mod webserver;
+
+extern crate actix_cors;
+extern crate actix_web;
+extern crate bellman;
+extern crate bls12_381;
+extern crate indexed_merkle_tree;
+extern crate rand;
+extern crate serde;
+extern crate serde_json;
