+++ crates/da/src/da.rs
@@ -1,16 +1,11 @@
-use crate::{
-    consts::CHANNEL_BUFFER_SIZE,
-    error::{DAResult, DataAvailabilityError, DeimosResult, GeneralError},
-    utils::Signable,
-    zk_snark::{Bls12Proof, VerifyingKey},
-};
 use async_trait::async_trait;
 use celestia_rpc::{BlobClient, Client, HeaderClient};
 use celestia_types::{blob::GasPrice, nmt::Namespace, Blob};
+use deimos_error::{DAResult, DataAvailabilityError, DeimosResult, GeneralError};
+use deimos_types::Signable;
+use deimos_zk_snark::{Bls12Proof, VerifyingKey};
 use ed25519::Signature;
-use fs2::FileExt;
 use serde::{Deserialize, Serialize};
-use serde_json::{json, Value};
 use std::{
     self,
     fs::{File, OpenOptions},
@@ -18,13 +13,23 @@ use std::{
     str::FromStr,
     sync::Arc,
 };
-use tokio::{
-    sync::{
-        mpsc::{channel, Receiver, Sender},
-        Mutex,
-    },
-    task::spawn,
+use tokio::sync::{
+    mpsc::{channel, Receiver, Sender},
+    Mutex,
 };
+use tokio::task::spawn;
+
+/// CHANNEL_BUFFER_SIZE determines the default channel size.
+pub const CHANNEL_BUFFER_SIZE: usize = 5;
+
+#[async_trait]
+pub trait DataAvailabilityLayer: Send + Sync {
+    async fn get_message(&self) -> DAResult<u64>;
+    async fn initialize_sync_target(&self) -> DAResult<u64>;
+    async fn get(&self, height: u64) -> DAResult<Vec<EpochJson>>;
+    async fn submit(&self, epoch: &EpochJson) -> DAResult<u64>;
+    async fn start(&self) -> DAResult<()>;
+}
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct EpochJson {
@@ -74,22 +79,6 @@ impl Signable for EpochJson {
     }
 }
 
-#[async_trait]
-pub trait DataAvailabilityLayer: Send + Sync {
-    async fn get_message(&self) -> DAResult<u64>;
-    async fn initialize_sync_target(&self) -> DAResult<u64>;
-    async fn get(&self, height: u64) -> DAResult<Vec<EpochJson>>;
-    async fn submit(&self, epoch: &EpochJson) -> DAResult<u64>;
-    async fn start(&self) -> DAResult<()>;
-}
-
-pub struct CelestiaConnection {
-    pub client: celestia_rpc::Client,
-    pub namespace_id: Namespace,
-
-    synctarget_tx: Arc<Sender<u64>>,
-    synctarget_rx: Arc<Mutex<Receiver<u64>>>,
-}
 
 /// The `NoopDataAvailabilityLayer` is a mock implementation of the `DataAvailabilityLayer` trait.
 pub struct NoopDataAvailabilityLayer {}
@@ -123,170 +112,6 @@ impl DataAvailabilityLayer for NoopDataAvailabilityLayer {
 ///
 /// This implementation is intended for testing and development only and should not be used in production environments. It provides a way to test the interactions with the data availability layer without the overhead of real network communication or data persistence.
 pub struct LocalDataAvailabilityLayer {}
-
-impl CelestiaConnection {
-    // TODO: Should take config
-    pub async fn new(
-        connection_string: &String,
-        auth_token: Option<&str>,
-        namespace_hex: &String,
-    ) -> DAResult<Self> {
-        let (tx, rx) = channel(CHANNEL_BUFFER_SIZE);
-
-        let client = Client::new(&connection_string, auth_token)
-            .await
-            .map_err(|e| {
-                DataAvailabilityError::ConnectionError(format!(
-                    "websocket initialization failed: {}",
-                    e
-                ))
-            })?;
-
-        let decoded_hex = match hex::decode(namespace_hex) {
-            Ok(hex) => hex,
-            Err(e) => {
-                return Err(DataAvailabilityError::GeneralError(
-                    GeneralError::DecodingError(format!(
-                        "decoding namespace '{}': {}",
-                        namespace_hex, e
-                    )),
-                ))
-            }
-        };
-
-        let namespace_id = Namespace::new_v0(&decoded_hex).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::EncodingError(format!(
-                "creating namespace '{}': {}",
-                namespace_hex, e
-            )))
-        })?;
-
-        Ok(CelestiaConnection {
-            client,
-            namespace_id,
-            synctarget_tx: Arc::new(tx),
-            synctarget_rx: Arc::new(Mutex::new(rx)),
-        })
-    }
-}
-
-#[async_trait]
-impl DataAvailabilityLayer for CelestiaConnection {
-    async fn get_message(&self) -> DAResult<u64> {
-        match self.synctarget_rx.lock().await.recv().await {
-            Some(height) => Ok(height),
-            None => Err(DataAvailabilityError::ChannelReceiveError),
-        }
-    }
-
-    async fn initialize_sync_target(&self) -> DAResult<u64> {
-        match HeaderClient::header_network_head(&self.client).await {
-            Ok(extended_header) => Ok(extended_header.header.height.value()),
-            Err(err) => Err(DataAvailabilityError::NetworkError(format!(
-                "getting network head from da layer: {}",
-                err
-            ))),
-        }
-    }
-
-    async fn get(&self, height: u64) -> DAResult<Vec<EpochJson>> {
-        trace!("searching for epoch on da layer at height {}", height);
-        match BlobClient::blob_get_all(&self.client, height, &[self.namespace_id]).await {
-            Ok(blobs) => {
-                let mut epochs = Vec::new();
-                for blob in blobs.iter() {
-                    match EpochJson::try_from(blob) {
-                        Ok(epoch_json) => epochs.push(epoch_json),
-                        Err(_) => {
-                            DataAvailabilityError::GeneralError(GeneralError::ParsingError(
-                                format!(
-                                    "marshalling blob from height {} to epoch json: {}",
-                                    height,
-                                    serde_json::to_string(&blob).unwrap()
-                                ),
-                            ));
-                        }
-                    }
-                }
-                Ok(epochs)
-            }
-            Err(err) => Err(DataAvailabilityError::DataRetrievalError(
-                height,
-                format!("getting epoch from da layer: {}", err),
-            )),
-        }
-    }
-
-    async fn submit(&self, epoch: &EpochJson) -> DAResult<u64> {
-        debug!("posting epoch {} to da layer", epoch.height);
-
-        let data = serde_json::to_string(&epoch).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::ParsingError(format!(
-                "serializing epoch json: {}",
-                e
-            )))
-        })?;
-        let blob = Blob::new(self.namespace_id.clone(), data.into_bytes()).map_err(|e| {
-            DataAvailabilityError::GeneralError(GeneralError::BlobCreationError(e.to_string()))
-        })?;
-        debug!(
-            "submitted blob with commitment {:?}",
-            serde_json::to_string(&blob.clone().commitment).unwrap()
-        );
-        trace!("blob: {:?}", serde_json::to_string(&blob).unwrap());
-        match self
-            .client
-            .blob_submit(&[blob.clone()], GasPrice::from(-1.0))
-            .await
-        {
-            Ok(height) => Ok(height),
-            Err(err) => Err(DataAvailabilityError::SubmissionError(
-                epoch.height,
-                err.to_string(),
-            )),
-        }
-    }
-
-    async fn start(&self) -> DAResult<()> {
-        let mut header_sub = HeaderClient::header_subscribe(&self.client)
-            .await
-            .map_err(|e| {
-                DataAvailabilityError::NetworkError(format!(
-                    "subscribing to headers from da layer: {}",
-                    e
-                ))
-            })?;
-
-        let synctarget_buffer = self.synctarget_tx.clone();
-        spawn(async move {
-            while let Some(extended_header_result) = header_sub.next().await {
-                match extended_header_result {
-                    Ok(extended_header) => {
-                        let height = extended_header.header.height.value();
-                        match synctarget_buffer.send(height).await {
-                            Ok(_) => {
-                                debug!("sent sync target update for height {}", height);
-                            }
-                            Err(_) => {
-                                DataAvailabilityError::SyncTargetError(format!(
-                                    "sending sync target update message for height {}",
-                                    height
-                                ));
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        DataAvailabilityError::NetworkError(format!(
-                            "retrieving header from da layer: {}",
-                            e
-                        ));
-                    }
-                }
-            }
-        });
-        Ok(())
-    }
-}
 
 impl LocalDataAvailabilityLayer {
     pub fn new() -> Self {
@@ -376,7 +201,7 @@ impl DataAvailabilityLayer for LocalDataAvailabilityLayer {
 }
 
 #[cfg(test)]
-mod da_tests {
+mod tests {
     use crate::{
         utils::validate_epoch,
         zk_snark::{
