+++ crates/da/src/lib.rs
@@ -1,0 +1,15 @@
+pub mod config;
+pub mod da;
+pub mod celestia;
+
+extern crate async_trait;
+extern crate celestia_rpc;
+extern crate celestia_types;
+extern crate deimos_errors;
+extern crate deimos_zk_snark;
+extern crate ed25519;
+extern crate serde;
+extern crate tokio;
+
+#[macro_use]
+extern crate log;
