+++ crates/errors/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod errors;
+
+extern crate indexed_merkle_tree;
+extern crate thiserror;
