+++ crates/node/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod lightclient;
+pub mod node;
+pub mod sequencer;
+mod test;
