+++ crates/storage/src/utils.rs
@@ -1,0 +1,28 @@
+use deimos_types::types::{ChainEntry, Operation};
+use indexed_merkle_tree::tree::Proof;
+
+/// Checks if a given public key in the list of `ChainEntry` objects has been revoked.
+///
+/// # Arguments
+///
+/// * `entries` - list of `ChainEntry` objects to be searched.
+/// * `value` - The value (public key) to be checked.
+///
+/// # Returns
+///
+/// `true` if the value was not revoked, otherwise `false`.
+/// TODO(@distractedm1nd): is_revoked > is_not_revoked, for readability
+pub fn is_not_revoked(entries: &[ChainEntry], value: String) -> bool {
+    for entry in entries {
+        if entry.value == value && matches!(entry.operation, Operation::Revoke) {
+            return false;
+        }
+    }
+    true
+}
+
+pub fn parse_json_to_proof(json_str: &str) -> Result<Proof, Box<dyn std::error::Error>> {
+    let proof: Proof = serde_json::from_str(json_str)?;
+
+    Ok(proof)
+}
