+++ crates/types/Cargo.toml
@@ -1,19 +1,13 @@
 [package]
-name = "deimos-storage"
+name = "deimos-types"
 
 [dependencies]
-deimos-types.workspace = true
 deimos-errors.workspace = true
 base64.workspace = true
 redis.workspace = true
 async-redis-session.workspace = true
-indexed-merkle-tree.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 ed25519-dalek.workspace = true
 ed25519.workspace = true
 mockall.workspace = true
-log.workspace = true
-
-[dev-dependencies]
-mockall = "0.12.1"
