+++ crates/zk_snark/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "deimos-zk-snark"
+
+[dependencies]
+log.workspace = true
+deimos-errors.workspace = true
+deimos-types.workspace = true
+base64.workspace = true
+bellman.workspace = true
+bls12_381.workspace = true
+indexed-merkle-tree.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+rand.workspace = true
+hex.workspace = true
