+++ crates/zk_snark/src/lib.rs
@@ -1,17 +1,15 @@
+pub mod zk_snark;
+pub mod utils;
+
 extern crate base64;
 extern crate deimos_errors;
 extern crate deimos_types;
-extern crate ed25519;
-extern crate ed25519_dalek;
+extern crate bellman;
+extern crate bls12_381;
+extern crate serde_json;
 extern crate indexed_merkle_tree;
-extern crate mockall;
-extern crate redis;
 extern crate serde;
-
-pub mod config;
-pub mod redis_db;
-pub mod storage;
-pub mod utils;
+extern crate rand;
 
 #[macro_use]
 extern crate log;
