+++ crates/zk_snark/src/utils.rs
@@ -1,57 +1,12 @@
-use crate::{
-    error::{DeimosError, DeimosResult, GeneralError, ProofError},
-    storage::{ChainEntry, Operation},
-    zk_snark::{
-        hex_to_scalar, InsertMerkleProofCircuit, ProofVariantCircuit, UpdateMerkleProofCircuit,
-    },
+use deimos_errors::errors::{DeimosError, DeimosResult, GeneralError,ProofError};
+use crate::zk_snark::{
+    hex_to_scalar, InsertMerkleProofCircuit, UpdateMerkleProofCircuit
 };
-use base64::{engine::general_purpose::STANDARD as engine, Engine as _};
 use bellman::groth16::{self, VerifyingKey};
 use bls12_381::{Bls12, Scalar};
-use ed25519::Signature;
-use ed25519_dalek::{Verifier, VerifyingKey as Ed25519VerifyingKey};
-use indexed_merkle_tree::tree::{InsertProof, NonMembershipProof, Proof, UpdateProof};
-use rand::rngs::OsRng;
+use indexed_merkle_tree::tree::{InsertProof, NonMembershipProof, UpdateProof};
+use serde_json::{self};
 
-/// Checks if a given public key in the list of `ChainEntry` objects has been revoked.
-///
-/// # Arguments
-///
-/// * `entries` - list of `ChainEntry` objects to be searched.
-/// * `value` - The value (public key) to be checked.
-///
-/// # Returns
-///
-/// `true` if the value was not revoked, otherwise `false`.
-/// TODO(@distractedm1nd): is_revoked > is_not_revoked, for readability
-pub fn is_not_revoked(entries: &[ChainEntry], value: String) -> bool {
-    for entry in entries {
-        if entry.value == value && matches!(entry.operation, Operation::Revoke) {
-            return false;
-        }
-    }
-    true
-}
-
-pub fn parse_json_to_proof(json_str: &str) -> Result<Proof, Box<dyn std::error::Error>> {
-    let proof: Proof = serde_json::from_str(json_str)?;
-
-    Ok(proof)
-}
-
-pub fn decode_public_key(pub_key_str: &String) -> DeimosResult<Ed25519VerifyingKey> {
-    // decode the public key from base64 string to bytes
-    let public_key_bytes = engine
-        .decode(pub_key_str)
-        .map_err(|e| GeneralError::DecodingError(format!("hex string: {}", e)))?;
-
-    let public_key_array: [u8; 32] = public_key_bytes
-        .try_into()
-        .map_err(|_| GeneralError::ParsingError("Vec<u8> to [u8; 32]".to_string()))?;
-
-    Ed25519VerifyingKey::from_bytes(&public_key_array)
-        .map_err(|_| GeneralError::DecodingError("ed25519 verifying key".to_string()).into())
-}
 
 pub fn validate_proof(proof_value: String) -> DeimosResult<()> {
     if let Ok((non_membership_proof, first_proof, second_proof)) =
@@ -89,34 +44,6 @@ pub fn validate_proof(proof_value: String) -> DeimosResult<()> {
     }
 }
 
-pub fn create_and_verify_snark(
-    circuit: ProofVariantCircuit,
-    scalars: Vec<Scalar>,
-) -> DeimosResult<(groth16::Proof<Bls12>, VerifyingKey<Bls12>)> {
-    let rng = &mut OsRng;
-
-    trace!("creating parameters with BLS12-381 pairing-friendly elliptic curve construction....");
-    let params =
-        groth16::generate_random_parameters::<Bls12, _, _>(circuit.clone(), rng).map_err(|e| {
-            DeimosError::Proof(ProofError::ProofUnpackError(format!(
-                "generating random params: {}",
-                e
-            )))
-        })?;
-
-    trace!("creating proof for zkSNARK...");
-    let proof = groth16::create_random_proof(circuit, &params, rng)
-        .map_err(|e| DeimosError::Proof(ProofError::GenerationError(e.to_string())))?;
-
-    trace!("preparing verifying key for zkSNARK...");
-    let pvk = groth16::prepare_verifying_key(&params.vk);
-
-    groth16::verify_proof(&pvk, &proof, &scalars)
-        .map_err(|e| DeimosError::Proof(ProofError::VerificationError(e.to_string())))?;
-
-    Ok((proof, params.vk))
-}
-
 pub fn validate_epoch(
     previous_commitment: &String,
     current_commitment: &String,
@@ -151,40 +78,13 @@ pub fn validate_epoch(
     Ok(proof)
 }
 
-pub trait Signable {
-    fn get_signature(&self) -> DeimosResult<Signature>;
-    fn get_content_to_sign(&self) -> DeimosResult<String>;
-    fn get_public_key(&self) -> DeimosResult<String>;
-}
-
-// verifies the signature of a given signable item and returns the content of the item if the signature is valid
-pub fn verify_signature<T: Signable>(
-    item: &T,
-    optional_public_key: Option<String>,
-) -> DeimosResult<String> {
-    let public_key_str = match optional_public_key {
-        Some(key) => key,
-        None => item.get_public_key()?,
-    };
-
-    let public_key = decode_public_key(&public_key_str)
-        .map_err(|_| DeimosError::General(GeneralError::InvalidPublicKey))?;
-
-    let content = item.get_content_to_sign()?;
-    let signature = item.get_signature()?;
-
-    if public_key.verify(content.as_bytes(), &signature).is_ok() {
-        Ok(content)
-    } else {
-        Err(GeneralError::InvalidSignature.into())
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use rand::rngs::OsRng;
     use crate::zk_snark::BatchMerkleProofCircuit;
     use indexed_merkle_tree::tree::{IndexedMerkleTree, Proof};
 
+    use deimos_types::types::decode_public_key;
     use indexed_merkle_tree::{node::Node, sha256};
 
     use super::*;
