+++ crates/zk_snark/src/zk_snark.rs
@@ -1,8 +1,5 @@
-use crate::{
-    error::{DeimosError, GeneralError, ProofError},
-    storage::ChainEntry,
-    utils::create_and_verify_snark,
-};
+use deimos_errors::errors::{DeimosResult, DeimosError, GeneralError, ProofError};
+use deimos_types::types::ChainEntry;
 use base64::{engine::general_purpose::STANDARD as engine, Engine as _};
 use bellman::{groth16, Circuit, ConstraintSystem, SynthesisError};
 use bls12_381::{Bls12, G1Affine, G2Affine, Scalar};
@@ -12,6 +9,8 @@ use indexed_merkle_tree::{
     tree::{InsertProof, MerkleProof, Proof, UpdateProof},
 };
 use serde::{Deserialize, Serialize};
+use rand::rngs::OsRng;
+
 
 fn vec_to_96_array(vec: Vec<u8>) -> Result<[u8; 96], DeimosError> {
     let mut array = [0u8; 96];
@@ -33,6 +32,34 @@ fn vec_to_192_array(vec: Vec<u8>) -> Result<[u8; 192], DeimosError> {
     }
     array.copy_from_slice(&vec);
     Ok(array)
+}
+
+pub fn create_and_verify_snark(
+    circuit: ProofVariantCircuit,
+    scalars: Vec<Scalar>,
+) -> DeimosResult<(groth16::Proof<Bls12>, groth16::VerifyingKey<Bls12>)> {
+    let rng = &mut OsRng;
+
+    trace!("creating parameters with BLS12-381 pairing-friendly elliptic curve construction....");
+    let params =
+        groth16::generate_random_parameters::<Bls12, _, _>(circuit.clone(), rng).map_err(|e| {
+            DeimosError::Proof(ProofError::ProofUnpackError(format!(
+                "generating random params: {}",
+                e
+            )))
+        })?;
+
+    trace!("creating proof for zkSNARK...");
+    let proof = groth16::create_random_proof(circuit, &params, rng)
+        .map_err(|e| DeimosError::Proof(ProofError::GenerationError(e.to_string())))?;
+
+    trace!("preparing verifying key for zkSNARK...");
+    let pvk = groth16::prepare_verifying_key(&params.vk);
+
+    groth16::verify_proof(&pvk, &proof, &scalars)
+        .map_err(|e| DeimosError::Proof(ProofError::VerificationError(e.to_string())))?;
+
+    Ok((proof, params.vk))
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
@@ -167,104 +194,6 @@ pub fn deserialize_custom_to_verifying_key(
         delta_g2,
         ic,
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::zk_snark::deserialize_proof;
-
-    use super::*;
-    use bellman::groth16;
-    use bls12_381::Bls12;
-    use indexed_merkle_tree::{node::Node, sha256, tree::IndexedMerkleTree};
-    use rand::rngs::OsRng;
-
-    const EMPTY_HASH: &str = Node::EMPTY_HASH;
-    const TAIL: &str = Node::TAIL;
-
-    fn build_empty_tree() -> IndexedMerkleTree {
-        let active_node = Node::new_leaf(
-            true,
-            true,
-            EMPTY_HASH.to_string(),
-            EMPTY_HASH.to_string(),
-            TAIL.to_string(),
-        );
-        let inactive_node = Node::new_leaf(
-            false,
-            true,
-            EMPTY_HASH.to_string(),
-            EMPTY_HASH.to_string(),
-            TAIL.to_string(),
-        );
-
-        // build a tree with 4 nodes
-        IndexedMerkleTree::new(vec![
-            active_node,
-            inactive_node.clone(),
-            inactive_node.clone(),
-            inactive_node,
-        ])
-        .unwrap()
-    }
-
-    #[test]
-    fn test_serialize_and_deserialize_proof() {
-        let mut tree = build_empty_tree();
-        let prev_commitment = tree.get_commitment().unwrap();
-
-        // create two nodes to insert
-        let ryan = sha256(&"Ryan".to_string());
-        let ford = sha256(&"Ford".to_string());
-        let sebastian = sha256(&"Sebastian".to_string());
-        let pusch = sha256(&"Pusch".to_string());
-        let ryans_node = Node::new_leaf(true, true, ryan, ford, TAIL.to_string());
-        let sebastians_node = Node::new_leaf(true, true, sebastian, pusch, TAIL.to_string());
-
-        // generate proofs for the two nodes
-        let first_insert_proof = tree.insert_node(&ryans_node).unwrap();
-        let second_insert_proof = tree.insert_node(&sebastians_node).unwrap();
-
-        // create zkSNARKs for the two proofs
-        let first_insert_zk_snark = Proof::Insert(first_insert_proof);
-        let second_insert_zk_snark = Proof::Insert(second_insert_proof);
-
-        let proofs = vec![first_insert_zk_snark, second_insert_zk_snark];
-        let current_commitment = tree.get_commitment().unwrap();
-
-        let batched_proof =
-            BatchMerkleProofCircuit::new(&prev_commitment, &current_commitment, proofs).unwrap();
-
-        let rng = &mut OsRng;
-        let params =
-            groth16::generate_random_parameters::<Bls12, _, _>(batched_proof.clone(), rng).unwrap();
-        let proof = groth16::create_random_proof(batched_proof.clone(), &params, rng).unwrap();
-
-        let serialized_proof = serialize_proof(&proof);
-        let deserialized_proof_result = deserialize_proof(&serialized_proof);
-        assert!(deserialized_proof_result.is_ok(), "Deserialization failed");
-
-        let deserialized_proof = deserialized_proof_result.unwrap();
-        assert_eq!(proof.a, deserialized_proof.a);
-        assert_eq!(proof.b, deserialized_proof.b);
-        assert_eq!(proof.c, deserialized_proof.c);
-    }
-
-    #[test]
-    fn test_deserialize_invalid_proof() {
-        // Erstellen Sie ein ungültiges Bls12Proof-Objekt
-        let invalid_proof = Bls12Proof {
-            a: "blubbubs".to_string(),
-            b: "FTV0oqNyecdbzY9QFPe5gfiQbSn1E0t+QHn+l+Ey6G2Dk0UZFm1wMsnRbIp5HCneDC+jf6rHCADL1NQ9FIF9o5Td8jObATCRm/YoIoeXY1yFY1rCEoJWFZU0zPeOR7XfBEmccqdMATwb8yznOj6Hn9XqZIr7E3C0XBtzk9GiahLopjP+SN9v/KLEpnLm3dn5FeAp7TcJ0gibi4nNT3u2vziKRNiDIKl71bp6tNC6grCdGOazpkrFSxiYi3QHJOYI".to_string(),
-            c: "BEKZboEyoJ3l+DLIF8IMjUR2kJQ9aq2kuXTZR8YizcQMg7zTH0xLO9JtTueneS3JFx1KlK6e2NkFZamiQERujx6bhmwIDgY8ZPCJ8iG//4E3eS0CZ25CJfnOucLeotyr".to_string(),
-        };
-
-        // Versuchen Sie, das ungültige Objekt zu deserialisieren
-        let deserialized_proof_result = deserialize_proof(&invalid_proof);
-
-        // Überprüfen Sie, ob die Deserialisierung fehlgeschlagen ist
-        assert!(deserialized_proof_result.is_err());
-    }
 }
 
 #[derive(Clone)]
@@ -735,5 +664,103 @@ impl HashChainEntryCircuit {
     pub fn create_public_parameter(value: &String) -> Result<Scalar, GeneralError> {
         let hashed_value = sha256(&value);
         Ok(hex_to_scalar(&hashed_value)?)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::zk_snark::deserialize_proof;
+
+    use super::*;
+    use bellman::groth16;
+    use bls12_381::Bls12;
+    use indexed_merkle_tree::{node::Node, sha256, tree::IndexedMerkleTree};
+    use rand::rngs::OsRng;
+
+    const EMPTY_HASH: &str = Node::EMPTY_HASH;
+    const TAIL: &str = Node::TAIL;
+
+    fn build_empty_tree() -> IndexedMerkleTree {
+        let active_node = Node::new_leaf(
+            true,
+            true,
+            EMPTY_HASH.to_string(),
+            EMPTY_HASH.to_string(),
+            TAIL.to_string(),
+        );
+        let inactive_node = Node::new_leaf(
+            false,
+            true,
+            EMPTY_HASH.to_string(),
+            EMPTY_HASH.to_string(),
+            TAIL.to_string(),
+        );
+
+        // build a tree with 4 nodes
+        IndexedMerkleTree::new(vec![
+            active_node,
+            inactive_node.clone(),
+            inactive_node.clone(),
+            inactive_node,
+        ])
+        .unwrap()
+    }
+
+    #[test]
+    fn test_serialize_and_deserialize_proof() {
+        let mut tree = build_empty_tree();
+        let prev_commitment = tree.get_commitment().unwrap();
+
+        // create two nodes to insert
+        let ryan = sha256(&"Ryan".to_string());
+        let ford = sha256(&"Ford".to_string());
+        let sebastian = sha256(&"Sebastian".to_string());
+        let pusch = sha256(&"Pusch".to_string());
+        let ryans_node = Node::new_leaf(true, true, ryan, ford, TAIL.to_string());
+        let sebastians_node = Node::new_leaf(true, true, sebastian, pusch, TAIL.to_string());
+
+        // generate proofs for the two nodes
+        let first_insert_proof = tree.insert_node(&ryans_node).unwrap();
+        let second_insert_proof = tree.insert_node(&sebastians_node).unwrap();
+
+        // create zkSNARKs for the two proofs
+        let first_insert_zk_snark = Proof::Insert(first_insert_proof);
+        let second_insert_zk_snark = Proof::Insert(second_insert_proof);
+
+        let proofs = vec![first_insert_zk_snark, second_insert_zk_snark];
+        let current_commitment = tree.get_commitment().unwrap();
+
+        let batched_proof =
+            BatchMerkleProofCircuit::new(&prev_commitment, &current_commitment, proofs).unwrap();
+
+        let rng = &mut OsRng;
+        let params =
+            groth16::generate_random_parameters::<Bls12, _, _>(batched_proof.clone(), rng).unwrap();
+        let proof = groth16::create_random_proof(batched_proof.clone(), &params, rng).unwrap();
+
+        let serialized_proof = serialize_proof(&proof);
+        let deserialized_proof_result = deserialize_proof(&serialized_proof);
+        assert!(deserialized_proof_result.is_ok(), "Deserialization failed");
+
+        let deserialized_proof = deserialized_proof_result.unwrap();
+        assert_eq!(proof.a, deserialized_proof.a);
+        assert_eq!(proof.b, deserialized_proof.b);
+        assert_eq!(proof.c, deserialized_proof.c);
+    }
+
+    #[test]
+    fn test_deserialize_invalid_proof() {
+        // Erstellen Sie ein ungültiges Bls12Proof-Objekt
+        let invalid_proof = Bls12Proof {
+            a: "blubbubs".to_string(),
+            b: "FTV0oqNyecdbzY9QFPe5gfiQbSn1E0t+QHn+l+Ey6G2Dk0UZFm1wMsnRbIp5HCneDC+jf6rHCADL1NQ9FIF9o5Td8jObATCRm/YoIoeXY1yFY1rCEoJWFZU0zPeOR7XfBEmccqdMATwb8yznOj6Hn9XqZIr7E3C0XBtzk9GiahLopjP+SN9v/KLEpnLm3dn5FeAp7TcJ0gibi4nNT3u2vziKRNiDIKl71bp6tNC6grCdGOazpkrFSxiYi3QHJOYI".to_string(),
+            c: "BEKZboEyoJ3l+DLIF8IMjUR2kJQ9aq2kuXTZR8YizcQMg7zTH0xLO9JtTueneS3JFx1KlK6e2NkFZamiQERujx6bhmwIDgY8ZPCJ8iG//4E3eS0CZ25CJfnOucLeotyr".to_string(),
+        };
+
+        // Versuchen Sie, das ungültige Objekt zu deserialisieren
+        let deserialized_proof_result = deserialize_proof(&invalid_proof);
+
+        // Überprüfen Sie, ob die Deserialisierung fehlgeschlagen ist
+        assert!(deserialized_proof_result.is_err());
     }
 }
