--- src/lib.rs
@@ -1,10 +1,0 @@
-pub mod cfg;
-pub mod consts;
-pub mod da;
-pub mod error;
-pub mod node_types;
-pub mod utils;
-mod webserver;
-pub mod zk_snark;
-#[macro_use]
-extern crate log;
